backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
tiktoken==0.7.0
faiss-cpu==1.8.0.post1
openai==1.44.0
numpy==1.26.4
pypdf==4.3.1
python-docx==1.1.2